moved failed stats collection test to 'tests/stats'
