+++ tests/stats/test_instrumentation.py
@@ -1,5 +1,9 @@
 import pytest
 
+from chatsky import Context
+from chatsky.core.service import Service, ExtraHandlerRuntimeInfo, ComponentExecutionState
+from tests.pipeline.utils import run_test_group
+
 try:
     from chatsky.stats import default_extractors
     from chatsky.stats import OtelInstrumentor
@@ -36,3 +40,15 @@ def test_keyword_arguments():
     assert instrumentor._meter_provider is not get_meter_provider()
     assert instrumentor._logger_provider is not get_logger_provider()
     assert instrumentor._tracer_provider is not get_tracer_provider()
+
+
+def test_failed_stats_collection():
+    chatsky_instrumentor = OtelInstrumentor.from_url("grpc://localhost:4317")
+    chatsky_instrumentor.instrument()
+
+    @chatsky_instrumentor
+    async def bad_stats_collector(_: Context, __: ExtraHandlerRuntimeInfo):
+        raise Exception
+
+    service = Service(handler=lambda _: None, before_handler=bad_stats_collector)
+    assert run_test_group(service) == ComponentExecutionState.FINISHED
